pyproject.toml
[project]
name = "invoice-matcher"
version = "1.0.0"
description = "Automatically reconcile bank statements with invoice PDFs"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "pdfplumber>=0.10.0",
    "pandas>=2.0.0",
    "rapidfuzz>=3.0.0",
    "tabulate>=0.9.0",
    "fiobank>=4.0.0",
    # Web dependencies
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    # Google Drive
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["models", "parsers", "matching", "reports"]

[tool.hatch.build.targets.wheel.sources]
"" = ""

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
